# Conflicts:
#	Source/utils/paths.cpp

+++ Source/loadsave.h
@@ -10,7 +10,7 @@
 namespace devilution {
 
 extern bool gbIsHellfireSaveGame;
-extern int giNumberOfLevels;
+extern uint8_t giNumberOfLevels;
 
 void RemoveInvalidItem(ItemStruct *pItem);
 int RemapItemIdxFromDiablo(int i);
@@ -31,4 +31,4 @@ void SaveGame();
 void SaveLevel();
 void LoadLevel();
 
-}
+} // namespace devilution
